reorg func defs, split .h into 2
