+++ prototypes/util_env.h
@@ -1,29 +1,14 @@
-#ifndef _ENV_UTILS_
-#define _ENV_UTILS_
+#ifndef _UTIL_ENV_
+#define _UTIL_ENV_
 
 extern char **environ;
 
-typedef struct _link {
-	char *path;
-	struct _link *prev;
-	struct _link *next;
-} linked_path;
-
-linked_path *init_path_chain();
-
-char **stash_env(char **env);
-char **reset_env();
-void print_paths();
-void _print_env();
 void init_env(void);
+char **reset_env();
+void print_env();
 
-/* man getenv(3) */
 char *_getenv(const char *name);
+int _setenv(const char *name, const char *value, int overwrite);
+int _unsetenv(const char *name);
 
-/* man setenv(3) */
-int _setenv( const char *name, const char *value, int overwrite);
-
-/* man unsetenv(3) */
-int _unsetenv( const char *name);
-
-#endif /* _ENV_UTILS_ */
+#endif /* _UTIL_ENV_ */
