+++ prototypes/util_path.c
@@ -1,9 +1,9 @@
 #include "util_str.h"
-#include "util_env.h"
+#include "util_path.h"
 #include <stdio.h>
 #include <stdlib.h>
 
-extern char **environ;
+/******** BASIC PATH HANDLING ********/
 
 char *navigate_path()
 {
@@ -36,6 +36,20 @@ char *navigate_path()
 	}
 }
 
+void print_paths()
+{
+	char *path = navigate_path();
+
+	do {
+		printf("%s\n", path);
+		path = navigate_path();
+	} while (path);
+}
+
+/******** PATH CHAIN ********/
+
+/* comment out section if unneeded */
+
 linked_path *init_path_chain()
 {
 	linked_path
@@ -62,7 +76,6 @@ linked_path *init_path_chain()
 
 	return (head);
 }
-
 
 linked_path *nav_path_chain(linked_path *head)
 {
@@ -93,31 +106,3 @@ void erase_path_chain(linked_path *head)
 
 	free(prev);
 }
-
-void print_paths()
-{
-	char *path = navigate_path();
-
-	do {
-		printf("%s\n", path);
-		path = navigate_path();
-	} while (path);
-}
-
-int comp_addr(void **alpha, void **beta)
-{
-	printf("%p ?= %p", alpha, beta);
-
-	if (alpha == beta)
-		return (0);
-
-	return (-1);
-}
-
-/* void env_environ_are_the_same(char **env) */
-/* { */
-/* 	if (comp_addr((void **) env, (void **) environ) < 0) */
-/* 		printf(" are not the same\n"); */
-/* 	else */
-/* 		printf(" are equal\n"); */
-/* } */
